pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lilycloudproto"
version = "0.1.0"
description = "Lily Cloud Prototype"
authors = [
    { name = "2edbef4a9b", email = "echofedn@outlook.com" },
    { name = "oC10H15No", email = "richardyangruichen@gmail.com" },
    { name = "apple-37", email = "1833346463@qq.com" },
]
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.119.1",
    "pyjwt>=2.10.1",
    "pydantic-settings>=2.12.0",
    "pwdlib[argon2]>=0.3.0",
    "python-magic-bin>=0.4.14 ; sys_platform == 'win32'",
    "python-magic>=0.4.27 ; sys_platform != 'win32'",
    "scalar_fastapi>=1.4.3",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
requires-python = ">=3.13"

[project.optional-dependencies]
dev = ["black>=25.9.0", "mypy>=1.18.2", "pre-commit>=4.3.0", "ruff>=0.14.1"]

[tool.pyright]
venvPath = "."
venv = ".venv"
reportCallInDefaultInitializer = false

[tool.black]
target-version = ["py313"]
exclude = '''
/(
    \.venv
)/
'''

[tool.ruff]
target-version = "py313"
lint.select = ["E", "F", "B", "I", "UP", "C4", "SIM", "PL", "RUF"]
lint.ignore = ["B008"]

[tool.mypy]
python_version = "3.13"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
strict = true
